Updated readme with new run command
Web app was crashing due to a timeout error from the server. It was happenning because of a long waiitng time on riot server. I have established a waiting time for this which will keep the connection open for 5 minutes or as long as u need. It will certainly work out.